// ranger-app/src/models/scenario/scenario.ts
import {type Metric} from 'web-vitals';
import {type Capability} from './capability';
import {type Condition} from './condition';
import {type Entity} from './entity';
import {type Evaluation} from './evaluation';
import {type Feature} from './feature';
import {type Goal} from './goal';
import {type InfraNode} from './infrastructure';
import {type Inject} from './inject';
import {type Script} from './script';
import {type Story} from './story';
import {type TrainingLearningObjective} from './tlo';
import {type Vulnerability} from './vulnerability';

export type Scenario = {
  name: string;
  description?: string;
  start: string;
  end: string;
  nodes?: Record<string, Node>;
  features?: Record<string, Feature>;
  infrastructure?: Record<string, InfraNode>;
  conditions?: Record<string, Condition>;
  vulnerabilities?: Record<string, Vulnerability>;
  capabilities?: Record<string, Capability>;
  metrics?: Record<string, Metric>;
  evaluations?: Record<string, Evaluation>;
  tlos?: Record<string, TrainingLearningObjective>;
  entities?: Record<string, Entity>;
  goals?: Record<string, Goal>;
  injects?: Record<string, Inject>;
  events?: Record<string, Event>;
  scripts?: Record<string, Script>;
  stories?: Record<string, Story>;
};